requirements.txt
transformers==4.36.0
torch==2.1.0
gradio==4.8.0
httpx==0.25.2
loguru==0.7.2
python-dotenv==1.0.0
rich==13.7.0
requests==2.31.0